move unit tests into source code add 2 more unit tests add space to early exit
