+++ src/parser/parse_from_text/hashtag_content_char_ranges.rs
@@ -867,12 +867,12 @@ const HASHTAG_CONTENT_CHAR_RANGES: [RangeInclusive<u32>; NUMBER_OF_RANGES] = [
 ];
 
 #[derive(Debug, PartialEq, Eq)]
-pub enum FindRangeResult<'a> {
+enum FindRangeResult<'a> {
     WasOnRangeStart,
     Range(&'a RangeInclusive<u32>),
 }
 
-pub fn find_range_for_char<'a>(code: u32) -> FindRangeResult<'a> {
+fn find_range_for_char<'a>(code: u32) -> FindRangeResult<'a> {
     let index = HASHTAG_CONTENT_CHAR_RANGES.binary_search_by_key(&code, |range| *range.start());
     match index {
         Ok(_) => FindRangeResult::WasOnRangeStart,
@@ -884,5 +884,75 @@ pub fn find_range_for_char<'a>(code: u32) -> FindRangeResult<'a> {
             #[allow(clippy::integer_arithmetic, clippy::indexing_slicing)]
             index => FindRangeResult::Range(&HASHTAG_CONTENT_CHAR_RANGES[index - 1]),
         },
+    }
+}
+
+pub(crate) fn hashtag_content_char(c: char) -> bool {
+    if matches!(c, '#' | '﹟' | '＃' | ' ') {
+        false
+    } else if matches!(c, '+' | '-' | '_') {
+        true
+    } else {
+        let code: u32 = c as u32;
+        match find_range_for_char(code) {
+            FindRangeResult::WasOnRangeStart => true,
+            FindRangeResult::Range(range) => range.contains(&code),
+        }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use crate::parser::parse_from_text::hashtag_content_char_ranges::hashtag_content_char;
+
+    use super::{find_range_for_char, FindRangeResult, RangeInclusive};
+
+    #[test]
+    fn test_range_function() {
+        // these must return WasOnRangeStart
+        let codes: Vec<u32> = vec![0x30000, 0xe0100, 0x23, 0x30, 0x171f, 0x176e, 0x10fb0];
+        for code in codes.iter() {
+            assert_eq!(find_range_for_char(*code), FindRangeResult::WasOnRangeStart);
+        }
+
+        // these must be return associated ranges
+        let codes: Vec<(u32, RangeInclusive<u32>)> = vec![
+            (0x11066 + 5, 0x11066..=0x11075),  // in range
+            (0x11000 + 10, 0x11000..=0x11046), // in range
+            (0x11046 + 2, 0x11000..=0x11046),  // out of range
+            (0x10, 0x23..=0x23),
+            (0x09, 0x23..=0x23),
+            (0x0, 0x23..=0x23),
+            (0x25, 0x23..=0x23),
+            (0x2a + 1, 0x2a..=0x2a),
+            (0xfffff, 0xe0100..=0xe01ef),
+            // ^ this is beyond ranges and must return the
+            // last range
+        ];
+
+        for (code, range) in codes.iter() {
+            assert_eq!(find_range_for_char(*code), FindRangeResult::Range(range));
+        }
+    }
+
+    #[test]
+    fn test_some_valid_chars() -> Result<(), String> {
+        let valid_char_set = vec!['a', '+', '-', '_', 'ü', 'ä', 'ö', 'ß'];
+        for valid_char in valid_char_set {
+            if !hashtag_content_char(valid_char) {
+                return Err(format!("{} is valid but returned false", valid_char));
+            }
+        }
+        Ok(())
+    }
+    #[test]
+    fn test_some_invalid_chars() -> Result<(), String> {
+        let invalid_char_set = vec!['#', ' '];
+        for invalid_char in invalid_char_set {
+            if hashtag_content_char(invalid_char) {
+                return Err(format!("{} is invalid but returned true", invalid_char));
+            }
+        }
+        Ok(())
     }
 }
